(chore) Move from node-sass to dart-sass, use Bootstrap 4 variables
node-sass has been deprecated and does not support the correct usage
of the @use directive, so we need to update to dart-sass

The "Paper Dashboard" template is based on bootstrap 3. It would be
better to slowly migrate to a newer version. This commit already
introduces some of the variables/functions defined by Bootstrap4