Add double-entry accounting model for ethereum accounts
This change also removes the model that tried to keep track of account
balances via the database. The idea is that we are not going to
control the balances by tracking transactions, but only keep a notion
of the assets and liabilities that users and hub operators are
interested in.

This change also has a beneficial side-effect: the implementation of
methods that selects the best account and execution method are now
relegated to the client module.

Blockchain transfer execution create accounting transaction between
wallet and blockchain account
Canceled/failed transfers generate refund entry in accounting books

Add Transaction.gas_fee property
Blockchain transfers generate book entries related to gas fees

Raiden payments generate book entries