Simplified application initialization
The application was getting all the data at first and then refreshing
periodically. With the inclusion of multi-chain support, the amount of
data that the application will be dealing with might make this
approach too heavy. As an example, we can not load all the registered
tokens at the instance.

This commit removes all the initial store data instantiation and
pushes to the components that use it.