Support for user token management
Every user can now manage their own list of tokens that they would
like to use on the hub. As long as it is part of the chains supported
by the hub, the user should be able to accept payments and make
transfers

This commit adds another section to the main page, called 'Settings'.
At the moment, the only item to be managed there is this list of
users, but eventually we might add other sections there.